First draft of AI container

+++ 2-terminal-setup/starship.sh
@@ -2,16 +2,21 @@
 
 set -e
 
-pushd /tmp > /dev/null
-  export BIN_DIR=$HOME/.local/bin
-  mkdir --parents $BIN_DIR
-  rm --force install.sh
-  curl --silent --show-error --remote-name https://starship.rs/install.sh
-  chmod +x install.sh
-  ./install.sh --yes > /dev/null
-popd > /dev/null
+if [[ "$(which starship | wc -l)" == "0" ]]
+then
+  echo "starship missing - installing to $HOME/.local/bin"
+  pushd /tmp > /dev/null
+    export BIN_DIR=$HOME/.local/bin
+    mkdir --parents $BIN_DIR
+    rm --force install.sh
+    curl --silent --show-error --remote-name https://starship.rs/install.sh
+    chmod +x install.sh
+    ./install.sh --yes > /dev/null
+  popd > /dev/null
+
+fi
 
 mkdir --parents $HOME/.config
 cp starship.toml $HOME/.config/starship.toml
 
-echo 'eval $($HOME/.local/bin/starship init bash)' >> $HOME/.bashrc
+echo 'eval $(starship init bash)' >> $HOME/.bashrc
